rewrite variable types in the database and location update in the function showWeather

+++ app/src/main/java/com/example/weatherapp/viewmodel/WeatherAppViewModel.kt
@@ -145,7 +145,7 @@ class WeatherAppViewModel : ViewModel() {
 
                 stationId = data.id
                 cityName = data.city_name
-                temperature = data.temp?.toInt()
+                temperature = data.temp?.toDouble()
             }
             return weather
         }
